imprime algo que no deberia

+++ Kernel/systemCalls.c
@@ -10,7 +10,7 @@ void systemCallDispatcher(uint64_t rdi, uint64_t rsi, uint64_t rdx, uint64_t rcx
             //print_string((char*)rsi);
             break;
         case 1:
-            print_integer((int)rsi);
+            print_integer(rsi);
             break;
         case 2:
             break;
